"Add signal handling to save settings on exit"
Introduced SIGTERM and SIGINT signal handling to ensure settings are saved before the application exits. Moved `save_settings_stores` and `load_settings_stores` logic to `Utils.py` for better modularity and code organization. Added logging to confirm settings are saved or loaded properly.
